pyproject.toml
[project]
name = "office-prospector"
version = "1.0.0"
description = "Identify boutique tax preparation offices for acquisition from IRS FOIA data"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "httpx>=0.27",
    "pandas>=2.0",
    "thefuzz[speedup]",
    "googlemaps>=4.10",
    "beautifulsoup4",
    "lxml",
    "python-dotenv",
    "pyyaml",
    "tqdm",
    "dnspython>=2.4",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio"]

[project.scripts]
office-prospector = "src.cli:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
